Extract tree helpers and session utils out of App.tsx
Moves the 10 pure/impure helper functions and their types from
src/tui/App.tsx into src/tui/tree-helpers.ts and src/tui/session-utils.ts,
reducing App.tsx by ~490 lines. Updates the three test files to import from
the new modules instead of App.
